TRAN-130: Refactored the code after review, fixed bug
