+++ src/pages/menuUpload/orderDetails/OrderDetailsPage.tsx
@@ -26,7 +26,7 @@ import { PostOrderLineDetailsRequestPayload } from 'stores/types/orderLineDetail
 import { getRequest } from 'utils/apiClient';
 import { DEFAULT_OFFSET, EMPTY_SEARCHER, FIRST_PAGE } from 'utils/consts';
 
-import OrderLineDetails from './OrderLineDetails';
+import OrderLineDetails from '../../../components/forms/orderDetails/OrderLineDetails';
 
 function OrderDetailsPage() {
   const [displayAddModal, setDisplayAddModal] = useState(false);
